# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "API Security Test Platform",
    "description": "針對 OWASP API Top 10 及 Log4j 弱點的 API 測試站台",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://apitest.ryantseng.work",
      "description": "Production server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root endpoint",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "API Security Test Platform. See /docs for API list."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api1/user/{user_id}": {
      "get": {
        "summary": "Broken Object Level Authorization",
        "tags": ["API1: Broken Object Level Authorization"],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api2/login": {
      "post": {
        "summary": "Broken User Authentication",
        "tags": ["API2: Broken User Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api2/profile": {
      "get": {
        "summary": "未驗證存取個人資料",
        "tags": ["API2: Broken User Authentication"],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "note": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api3/data": {
      "get": {
        "summary": "Excessive Data Exposure",
        "tags": ["API3: Excessive Data Exposure"],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "password_hash": {
                      "type": "string"
                    },
                    "ssn": {
                      "type": "string"
                    },
                    "credit_card": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api4/slow": {
      "get": {
        "summary": "Lack of Resources & Rate Limiting",
        "tags": ["API4: Lack of Resources & Rate Limiting"],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api5/admin": {
      "get": {
        "summary": "Broken Function Level Authorization",
        "tags": ["API5: Broken Function Level Authorization"],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api6/profile": {
      "post": {
        "summary": "Mass Assignment",
        "tags": ["API6: Mass Assignment"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api7/debug": {
      "get": {
        "summary": "Security Misconfiguration",
        "tags": ["API7: Security Misconfiguration"],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "debug": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api8/sqli": {
      "get": {
        "summary": "SQL Injection",
        "tags": ["API8: Injection (SQLi, XSS)"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string"
                    },
                    "result": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api8/xss": {
      "get": {
        "summary": "XSS",
        "tags": ["API8: Injection (SQLi, XSS)"],
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api9/old-api": {
      "get": {
        "summary": "Improper Assets Management",
        "tags": ["API9: Improper Assets Management"],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api10/no-log": {
      "get": {
        "summary": "Insufficient Logging & Monitoring",
        "tags": ["API10: Insufficient Logging & Monitoring"],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/log4j/test": {
      "get": {
        "summary": "Log4j (Log4Shell)",
        "tags": ["Log4j (Log4Shell)"],
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
} 